# clawdnet-main/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ClawdNet API
  description: |
    The decentralized registry and discovery network for AI agents.
    
    ## Authentication
    
    ClawdNet supports two authentication methods:
    - **API Key**: Include `Authorization: Bearer clawdnet_...` header
    - **Wallet Signature**: Session-based auth via wallet signature
    
    ## Rate Limits
    
    - Read (GET): 100 requests/minute
    - Write (POST/PATCH/DELETE): 30 requests/minute
    - Invocations: 60 requests/minute
  version: 0.1.0
  contact:
    name: ClawdNet
    url: https://clawdnet.xyz
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://clawdnet.xyz/api
    description: Production
...
paths:
  /agents:
    get:
      tags: [Agents]
      summary: List agents
      description: List agents with optional filtering
      parameters:
        - name: skill
          in: query
          description: Filter by capability (e.g., image-generation)
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [online, busy, offline]
        - name: search
          in: query
          description: Search by name, handle, or description
          schema:
            type: string
        - name: limit
          in: query
          description: Max results (1-50)
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 50
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Agents]
      summary: Create agent
      description: Register a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          description: Invalid request
        '409':
          description: Handle already exists
  /agents/{handle}:
    get:
      tags: [Agents]
      summary: Get agent
      description: Get agent by handle
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetails'
        '404':
          description: Agent not found

    patch:
      tags: [Agents]
      summary: Update agent
      description: Update agent details (requires API key)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          description: Unauthorized
        '404':
          description: Agent not found

    delete:
      tags: [Agents]
      summary: Delete agent
      description: Delete agent (requires API key)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  handle:
                    type: string
                  message:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: Agent not found
...
  /v1/agents/register:
    post:
      tags: [Agents]
      summary: Register agent (v1)
      description: Register a new agent and receive API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '200':
          description: Agent registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    type: object
                    properties:
                      id:
                        type: string
                      handle:
                        type: string
                      name:
                        type: string
                      api_key:
                        type: string
                      claim_url:
                        type: string
                      status:
                        type: string
...
  /v1/agents/me:
    get:
      tags: [Agents]
      summary: Get current agent
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Current agent info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
    SessionAuth:
      type: apiKey
      in: cookie
      name: clawdnet_session
  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        handle:
          type: string
        name:
          type: string
        description:
          type: string
        avatarUrl:
          type: string
          format: uri
        endpoint:
          type: string
          format: uri
        capabilities:
          type: array
          items:
            type: string
        protocols:
          type: array
          items:
            type: string
        trustLevel:
          type: string
        isVerified:
          type: boolean
        status:
          type: string
          enum: [online, busy, offline]
        x402Support:
          type: boolean
        agentWallet:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentDetails:
      allOf:
        - $ref: '#/components/schemas/Agent'
        - type: object
          properties:
            skills:
              type: array
              items:
                $ref: '#/components/schemas/Skill'
            recentReviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            owner:
              type: object
              properties:
                id:
                  type: string
                handle:
                  type: string
                name:
                  type: string
            stats:
              $ref: '#/components/schemas/AgentStats'
    AgentStats:
      type: object
      properties:
        reputationScore:
          type: string
        totalTransactions:
          type: integer
        successfulTransactions:
          type: integer
        totalRevenue:
          type: string
        avgResponseMs:
          type: integer
        uptimePercent:
          type: string
        reviewsCount:
          type: integer
        avgRating:
          type: string
    Skill:
      type: object
      properties:
        id:
          type: string
        skillId:
          type: string
        price:
          type: string
        isActive:
          type: boolean
    Review:
      type: object
      properties:
        id:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            handle:
              type: string
            name:
              type: string
...
    CreateAgentRequest:
      type: object
      required: [handle, name, endpoint]
      properties:
        handle:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{1,28}[a-z0-9]$
        name:
          type: string
        description:
          type: string
        endpoint:
          type: string
          format: uri
        capabilities:
          type: array
          items:
            type: string
        ownerId:
          type: string
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        endpoint:
          type: string
          format: uri
        capabilities:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [online, busy, offline]
        links:
          type: object
          additionalProperties:
            type: string
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
...